Reorganize 1 - RoboSharp.Extensions.Comparers
This is to make the api more organized, where the root is used to expose the core functional classes and interfaces, and move the implementation helpers are one level deeper.

Files/Classes Moved:
RoboSharp.Extensions.Comparers
- DirectoryPairEqualityComparer
- FilePairEqualityComparer
- FilePairSorter
- PairEqualityComparer
- StringEqualityComparer
